requirements.txt
flask
flask-cors
google-generativeai>=0.7.0
gunicorn
fastapi
uvicorn
python-dotenv
pydantic

